Added checkConfigurationSheet and related parameters
Added functionality to check the "sheet_names" column of the "Modules_and_forms" sheet against the names of other sheets in the workbook. If an expected sheet doesn't exist, the cell will be flagged red in the Modules_and_forms sheet of the output workbook. The column name (default "sheet_names" is configurable via --configuration-sheet-column parameter, and the sheet name (default "Modules_and_forms") is configurable via --configuration-sheet.