Text format checking added
Very basic text format checking: takes a list of expected non-linguistic characters, compares the counts of the base column and the current column, and appends a text formatting mismatch, which also displays the diff for each character type between cur and base column. Also, if text formatting is the ONLY issue in a cell/row, the highlighting will be yellow instead of red.

This is activated via the --format-check feature flag. The set of characters to test can be set by --format-check-characters or added to by --format-check-charcaters-add

Non-linguistic character default:
~`!@#$%^&*()_-+={[}]|\\:;\"'<,>.?/